Updating the logic of converting numbers into words
In the file `numberToWord.cpp `added the `Triplet` structure and new functions `parceRightTriplet` and `getTripletName' for processing numbers. Outdated functions such as `parseRightDigit', `getUnitName', `getTenName` and `getHundredName` have been removed, which improves the code structure.

In the file `testsNumberToWord.cpp New tests have been added to verify the operation of the updated code, including tests for large numbers, which increases the reliability of the program.